requirements.txt
langchain
langchain-community
langchain-core
pydantic
requests
python-dotenv
sseclient-py
fastmcp 